fix reset_consumed_energy() function to support several refills:
In the previous implementation we were loading
the current _consumed_mah value on a variable
that we were then subtracting from _consumed_mah
in order to report the consumed mah since reset
to the rest of the system. We were asigning this
directly so in case more than one refills happened
the consumed mah would not make sense, as we would
have  needed to add the current consumed mah value
each time we reset.

The above was the first approach planned, but it seems
we are not using the accumulated consumed fuel liters
value anyway for other than reporting the consumed
liters since refill, so instead of keeping the old
variable and subtract it from _consumed_mah in this
commit we just reset to 0 accumulated_consumed_fuel
liters, so the result will be the same, with less
variables.