AP_Batt_Generator: Change how remaining capacity is processed to allow fuel refills or partial refills:
This commit allows setting a different
initial fuel capacity other than 100%
and allows to use MAV_CMD_BATTERY_RESET
cmd to comunicate to the autopilot
a refill, full or partial.

Failsafe management:
Override update_failsafes() from base class
to manage when we calculate our own fuel percentage
within this class ( when generator doesn't provide
fuel_remaining_pct.

When we are not calculating the percentage here,
call the regular backend method instead, as it was
before