AP_Generator_Loweheiser: change comms timeout and EFI min voltage:
and change them to defines instead.

The timeout is causing the GCS to receive unknown
fuel level information, as the timeout flags
loweheiser backend as unhealthy, and it propagates
all its way up to battmonitor.

It seems loweheiser sends its packet at 1 hz, but
it gets a bit delayed very often, causing this issue.

Raising a bit the timeout should solve this issue.

Regarding the voltage, apparently first versions
of loweheiser were meant to run at 13V, but they
changed something and indicated the healthy voltage
range should be 11 - 13.5 now for the EFI voltage